updated the file structure of table items procssings composable 'useTableItems'
