+++ packages/bootstrap-vue-next/src/composables/index.ts
@@ -16,3 +16,4 @@ export {default as useId} from './useId'
 export {default as useModalManager} from './useModalManager'
 export {default as useSafeScrollLock} from './useSafeScrollLock'
 export {default as useStateClass} from './useStateClass'
+export {default as useTableItems} from './useTableItems'
